@@ -1,6 +1,6 @@
 name: extended_image
 description: Official extension image, support placeholder(loading)/ failed state, cache network, zoom/pan, photo view, slide out page, editor(crop,rotate,flip), painting etc.
-version: 9.1.0
+version: 10.0.1
 repository: https://github.com/fluttercandies/extended_image
 issue_tracker: https://github.com/fluttercandies/extended_image/issues
 topics:
@@ -11,11 +11,11 @@ topics:
   - slide-out
 
 environment:
-  sdk: '>=2.18.0 <4.0.0'
-  flutter: '>=3.16.0'
+  sdk: ">=3.7.0 <4.0.0"
+  flutter: ">=3.29.0"
 
 dependencies:
-  extended_image_library: ^4.0.1
+  extended_image_library: ^5.0.0
   flutter:
     sdk: flutter
   meta: ^1.7.0
