@@ -17,7 +17,7 @@ class FlatButtonWithIcon extends TextButton {
           focusNode: focusNode,
           style: textColor != null
               ? ButtonStyle(
-                  textStyle: MaterialStateProperty.all<TextStyle>(
+                  textStyle: WidgetStateProperty.all<TextStyle>(
                   TextStyle(color: textColor),
                 ))
               : null,
@@ -135,7 +135,7 @@ class CommonCircularProgressIndicator extends StatelessWidget {
   Widget build(BuildContext context) {
     return Container(
       alignment: Alignment.center,
-      color: Colors.grey.withOpacity(0.8),
+      color: Colors.grey.withValues(alpha: 0.8),
       child: CircularProgressIndicator(
         strokeWidth: 2.0,
         valueColor:
