@@ -1,16 +1,17 @@
-// ignore_for_file: overridden_fields
+// ignore_for_file: overridden_fields, unused_element_parameter
+
+import 'dart:math';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
+import 'package:flutter/scheduler.dart';
 
 part 'drag_gesture_recognizer.dart';
 part 'velocity_tracker.dart';
 
-enum _DragState {
-  ready,
-  possible,
-  accepted,
-}
+enum _DragState { ready, possible, accepted }
+
+enum _DragDirection { horizontal, vertical }
 
 /// Recognizes movement.
 ///
@@ -31,26 +32,22 @@ enum _DragState {
 ///  * [HorizontalDragGestureRecognizer], for left and right drags.
 ///  * [VerticalDragGestureRecognizer], for up and down drags.
 ///  * [PanGestureRecognizer], for drags that are not locked to a single axis.
-abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
+sealed class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
   /// Initialize the object.
-  ///
-  /// [dragStartBehavior] must not be null.
   ///
   /// {@macro flutter.gestures.GestureRecognizer.supportedDevices}
   _DragGestureRecognizer({
     super.debugOwner,
     this.dragStartBehavior = DragStartBehavior.start,
+    this.multitouchDragStrategy = MultitouchDragStrategy.latestPointer,
     this.velocityTrackerBuilder = _defaultBuilder,
-    // ignore: unused_element
     this.onlyAcceptDragOnThreshold = false,
     super.supportedDevices,
-    AllowedButtonsFilter? allowedButtonsFilter,
-  }) : super(
-            allowedButtonsFilter:
-                allowedButtonsFilter ?? _defaultButtonAcceptBehavior);
+    super.allowedButtonsFilter = _defaultButtonAcceptBehavior,
+  });
 
-  static _VelocityTracker _defaultBuilder(PointerEvent event) =>
-      _VelocityTracker.withKind(event.kind);
+  static VelocityTracker _defaultBuilder(PointerEvent event) =>
+      VelocityTracker.withKind(event.kind);
 
   // Accept the input if, and only if, [kPrimaryButton] is pressed.
   static bool _defaultButtonAcceptBehavior(int buttons) =>
@@ -66,7 +63,7 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
   /// no difference in behavior between the two settings.
   ///
   /// For more information about the gesture arena:
-  /// https://flutter.dev/docs/development/ui/advanced/gestures#gesture-disambiguation
+  /// https://flutter.dev/to/gesture-disambiguation
   ///
   /// By default, the drag start behavior is [DragStartBehavior.start].
   ///
@@ -81,6 +78,29 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
   /// instead set to [DragStartBehavior.start], [onStart] will be called with
   /// position (510.0, 500.0).
   DragStartBehavior dragStartBehavior;
+
+  /// {@template flutter.gestures.monodrag.DragGestureRecognizer.multitouchDragStrategy}
+  /// Configure the multi-finger drag strategy on multi-touch devices.
+  ///
+  /// If set to [MultitouchDragStrategy.latestPointer], the drag gesture recognizer
+  /// will only track the latest active (accepted by this recognizer) pointer, which
+  /// appears to be only one finger dragging.
+  ///
+  /// If set to [MultitouchDragStrategy.averageBoundaryPointers], all active
+  /// pointers will be tracked, and the result is computed from the boundary pointers.
+  ///
+  /// If set to [MultitouchDragStrategy.sumAllPointers],
+  /// all active pointers will be tracked together and the scrolling offset
+  /// is the sum of the offsets of all active pointers
+  /// {@endtemplate}
+  ///
+  /// By default, the strategy is [MultitouchDragStrategy.latestPointer].
+  ///
+  /// See also:
+  ///
+  ///  * [MultitouchDragStrategy], which defines several different drag strategies for
+  ///  multi-finger drag.
+  MultitouchDragStrategy multitouchDragStrategy;
 
   /// A pointer has contacted the screen with a primary button and might begin
   /// to move.
@@ -226,6 +246,14 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
   _DragState _state = _DragState.ready;
   late OffsetPair _initialPosition;
   late OffsetPair _pendingDragOffset;
+
+  /// The local and global offsets of the last pointer event received.
+  ///
+  /// It is used to create the [DragEndDetails], which provides information about
+  /// the end of a drag gesture.
+  OffsetPair get lastPosition => _lastPosition;
+  late OffsetPair _lastPosition;
+
   Duration? _lastPendingEventTimestamp;
 
   /// When asserts are enabled, returns the last tracked pending event timestamp
@@ -254,6 +282,7 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
   ///
   /// If drag is only allowed along a defined axis, this value may be negative to
   /// differentiate the direction of the drag.
+  double get globalDistanceMoved => _globalDistanceMoved;
   late double _globalDistanceMoved;
 
   /// Determines if a gesture is a fling or not based on velocity.
@@ -268,16 +297,56 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
   /// A fling calls its gesture end callback with a velocity, allowing the
   /// provider of the callback to respond by carrying the gesture forward with
   /// inertia, for example.
-  DragEndDetails? _considerFling(
-      VelocityEstimate estimate, PointerDeviceKind kind);
+  DragEndDetails? considerFling(
+    VelocityEstimate estimate,
+    PointerDeviceKind kind,
+  );
 
+  /// Returns the effective delta that should be considered for the incoming [delta].
+  ///
+  /// The delta received by an event might contain both the x and y components
+  /// greater than zero, and an one-axis drag recognizer only cares about one
+  /// of them.
+  ///
+  /// For example, a [VerticalDragGestureRecognizer], would return an [Offset]
+  /// with the x component set to 0.0, because it only cares about the y component.
   Offset _getDeltaForDetails(Offset delta);
+
+  /// Returns the value for the primary axis from the given [value].
+  ///
+  /// For example, a [VerticalDragGestureRecognizer] would return the y
+  /// component, while a [HorizontalDragGestureRecognizer] would return
+  /// the x component.
+  ///
+  /// Returns `null` if the recognizer does not have a primary axis.
   double? _getPrimaryValueFromOffset(Offset value);
-  bool _hasSufficientGlobalDistanceToAccept(
-      PointerDeviceKind pointerDeviceKind, double? deviceTouchSlop);
+
+  /// The axis (horizontal or vertical) corresponding to the primary drag direction.
+  ///
+  /// The [PanGestureRecognizer] returns null.
+  _DragDirection? _getPrimaryDragAxis() => null;
+
+  /// Whether the [globalDistanceMoved] is big enough to accept the gesture.
+  ///
+  /// If this method returns `true`, it means this recognizer should declare win in the gesture arena.
+  bool hasSufficientGlobalDistanceToAccept(
+    PointerDeviceKind pointerDeviceKind,
+    double? deviceTouchSlop,
+  );
   bool _hasDragThresholdBeenMet = false;
 
   final Map<int, VelocityTracker> _velocityTrackers = <int, VelocityTracker>{};
+
+  // The move delta of each pointer before the next frame.
+  //
+  // The key is the pointer ID. It is cleared whenever a new batch of pointer events is detected.
+  final Map<int, Offset> _moveDeltaBeforeFrame = <int, Offset>{};
+
+  // The timestamp of all events of the current frame.
+  //
+  // On a event with a different timestamp, the event is considered a new batch.
+  Duration? _frameTimeStamp;
+  Offset _lastUpdatedDeltaForPan = Offset.zero;
 
   @override
   bool isPointerAllowed(PointerEvent event) {
@@ -300,17 +369,23 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
 
   void _addPointer(PointerEvent event) {
     _velocityTrackers[event.pointer] = velocityTrackerBuilder(event);
-    if (_state == _DragState.ready) {
-      _state = _DragState.possible;
-      _initialPosition =
-          OffsetPair(global: event.position, local: event.localPosition);
-      _pendingDragOffset = OffsetPair.zero;
-      _globalDistanceMoved = 0.0;
-      _lastPendingEventTimestamp = event.timeStamp;
-      _lastTransform = event.transform;
-      _checkDown();
-    } else if (_state == _DragState.accepted) {
-      resolve(GestureDisposition.accepted);
+    switch (_state) {
+      case _DragState.ready:
+        _state = _DragState.possible;
+        _initialPosition = OffsetPair(
+          global: event.position,
+          local: event.localPosition,
+        );
+        _lastPosition = _initialPosition;
+        _pendingDragOffset = OffsetPair.zero;
+        _globalDistanceMoved = 0.0;
+        _lastPendingEventTimestamp = event.timeStamp;
+        _lastTransform = event.transform;
+        _checkDown();
+      case _DragState.possible:
+        break;
+      case _DragState.accepted:
+        resolve(GestureDisposition.accepted);
     }
   }
 
@@ -333,6 +408,238 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
     _addPointer(event);
   }
 
+  bool _shouldTrackMoveEvent(int pointer) {
+    final bool result;
+    switch (multitouchDragStrategy) {
+      case MultitouchDragStrategy.sumAllPointers:
+      case MultitouchDragStrategy.averageBoundaryPointers:
+        result = true;
+      case MultitouchDragStrategy.latestPointer:
+        result = _activePointer == null || pointer == _activePointer;
+    }
+    return result;
+  }
+
+  void _recordMoveDeltaForMultitouch(int pointer, Offset localDelta) {
+    if (multitouchDragStrategy !=
+        MultitouchDragStrategy.averageBoundaryPointers) {
+      assert(_frameTimeStamp == null);
+      assert(_moveDeltaBeforeFrame.isEmpty);
+      return;
+    }
+
+    assert(
+      _frameTimeStamp == SchedulerBinding.instance.currentSystemFrameTimeStamp,
+    );
+
+    if (_state != _DragState.accepted || localDelta == Offset.zero) {
+      return;
+    }
+
+    if (_moveDeltaBeforeFrame.containsKey(pointer)) {
+      final Offset offset = _moveDeltaBeforeFrame[pointer]!;
+      _moveDeltaBeforeFrame[pointer] = offset + localDelta;
+    } else {
+      _moveDeltaBeforeFrame[pointer] = localDelta;
+    }
+  }
+
+  double _getSumDelta({
+    required int pointer,
+    required bool positive,
+    required _DragDirection axis,
+  }) {
+    double sum = 0.0;
+
+    if (!_moveDeltaBeforeFrame.containsKey(pointer)) {
+      return sum;
+    }
+
+    final Offset offset = _moveDeltaBeforeFrame[pointer]!;
+    if (positive) {
+      if (axis == _DragDirection.vertical) {
+        sum = max(offset.dy, 0.0);
+      } else {
+        sum = max(offset.dx, 0.0);
+      }
+    } else {
+      if (axis == _DragDirection.vertical) {
+        sum = min(offset.dy, 0.0);
+      } else {
+        sum = min(offset.dx, 0.0);
+      }
+    }
+
+    return sum;
+  }
+
+  int? _getMaxSumDeltaPointer({
+    required bool positive,
+    required _DragDirection axis,
+  }) {
+    if (_moveDeltaBeforeFrame.isEmpty) {
+      return null;
+    }
+
+    int? ret;
+    double? max;
+    double sum;
+    for (final int pointer in _moveDeltaBeforeFrame.keys) {
+      sum = _getSumDelta(pointer: pointer, positive: positive, axis: axis);
+      if (ret == null) {
+        ret = pointer;
+        max = sum;
+      } else {
+        if (positive) {
+          if (sum > max!) {
+            ret = pointer;
+            max = sum;
+          }
+        } else {
+          if (sum < max!) {
+            ret = pointer;
+            max = sum;
+          }
+        }
+      }
+    }
+    assert(ret != null);
+    return ret;
+  }
+
+  Offset _resolveLocalDeltaForMultitouch(int pointer, Offset localDelta) {
+    if (multitouchDragStrategy !=
+        MultitouchDragStrategy.averageBoundaryPointers) {
+      if (_frameTimeStamp != null) {
+        _moveDeltaBeforeFrame.clear();
+        _frameTimeStamp = null;
+        _lastUpdatedDeltaForPan = Offset.zero;
+      }
+      return localDelta;
+    }
+
+    final Duration currentSystemFrameTimeStamp =
+        SchedulerBinding.instance.currentSystemFrameTimeStamp;
+    if (_frameTimeStamp != currentSystemFrameTimeStamp) {
+      _moveDeltaBeforeFrame.clear();
+      _lastUpdatedDeltaForPan = Offset.zero;
+      _frameTimeStamp = currentSystemFrameTimeStamp;
+    }
+
+    assert(
+      _frameTimeStamp == SchedulerBinding.instance.currentSystemFrameTimeStamp,
+    );
+
+    final _DragDirection? axis = _getPrimaryDragAxis();
+
+    if (_state != _DragState.accepted ||
+        localDelta == Offset.zero ||
+        (_moveDeltaBeforeFrame.isEmpty && axis != null)) {
+      return localDelta;
+    }
+
+    final double dx, dy;
+    if (axis == _DragDirection.horizontal) {
+      dx = _resolveDelta(
+        pointer: pointer,
+        axis: _DragDirection.horizontal,
+        localDelta: localDelta,
+      );
+      assert(dx.abs() <= localDelta.dx.abs());
+      dy = 0.0;
+    } else if (axis == _DragDirection.vertical) {
+      dx = 0.0;
+      dy = _resolveDelta(
+        pointer: pointer,
+        axis: _DragDirection.vertical,
+        localDelta: localDelta,
+      );
+      assert(dy.abs() <= localDelta.dy.abs());
+    } else {
+      final double averageX = _resolveDeltaForPanGesture(
+        axis: _DragDirection.horizontal,
+        localDelta: localDelta,
+      );
+      final double averageY = _resolveDeltaForPanGesture(
+        axis: _DragDirection.vertical,
+        localDelta: localDelta,
+      );
+      final Offset updatedDelta =
+          Offset(averageX, averageY) - _lastUpdatedDeltaForPan;
+      _lastUpdatedDeltaForPan = Offset(averageX, averageY);
+      dx = updatedDelta.dx;
+      dy = updatedDelta.dy;
+    }
+
+    return Offset(dx, dy);
+  }
+
+  double _resolveDelta({
+    required int pointer,
+    required _DragDirection axis,
+    required Offset localDelta,
+  }) {
+    final bool positive =
+        axis == _DragDirection.horizontal
+            ? localDelta.dx > 0
+            : localDelta.dy > 0;
+    final double delta =
+        axis == _DragDirection.horizontal ? localDelta.dx : localDelta.dy;
+    final int? maxSumDeltaPointer = _getMaxSumDeltaPointer(
+      positive: positive,
+      axis: axis,
+    );
+    assert(maxSumDeltaPointer != null);
+
+    if (maxSumDeltaPointer == pointer) {
+      return delta;
+    } else {
+      final double maxSumDelta = _getSumDelta(
+        pointer: maxSumDeltaPointer!,
+        positive: positive,
+        axis: axis,
+      );
+      final double curPointerSumDelta = _getSumDelta(
+        pointer: pointer,
+        positive: positive,
+        axis: axis,
+      );
+      if (positive) {
+        if (curPointerSumDelta + delta > maxSumDelta) {
+          return curPointerSumDelta + delta - maxSumDelta;
+        } else {
+          return 0.0;
+        }
+      } else {
+        if (curPointerSumDelta + delta < maxSumDelta) {
+          return curPointerSumDelta + delta - maxSumDelta;
+        } else {
+          return 0.0;
+        }
+      }
+    }
+  }
+
+  double _resolveDeltaForPanGesture({
+    required _DragDirection axis,
+    required Offset localDelta,
+  }) {
+    final double delta =
+        axis == _DragDirection.horizontal ? localDelta.dx : localDelta.dy;
+    final int pointerCount = _acceptedActivePointers.length;
+    assert(pointerCount >= 1);
+
+    double sum = delta;
+    for (final Offset offset in _moveDeltaBeforeFrame.values) {
+      if (axis == _DragDirection.horizontal) {
+        sum += offset.dx;
+      } else {
+        sum += offset.dy;
+      }
+    }
+    return sum / pointerCount;
+  }
+
   @override
   void handleEvent(PointerEvent event) {
     assert(_state != _DragState.ready);
@@ -341,79 +648,101 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
             event is PointerMoveEvent ||
             event is PointerPanZoomStartEvent ||
             event is PointerPanZoomUpdateEvent)) {
-      final VelocityTracker tracker = _velocityTrackers[event.pointer]!;
-      if (event is PointerPanZoomStartEvent) {
-        tracker.addPosition(event.timeStamp, Offset.zero);
-      } else if (event is PointerPanZoomUpdateEvent) {
-        tracker.addPosition(event.timeStamp, event.pan);
-      } else {
-        tracker.addPosition(event.timeStamp, event.localPosition);
-      }
+      final Offset position = switch (event) {
+        PointerPanZoomStartEvent() => Offset.zero,
+        PointerPanZoomUpdateEvent() => event.pan,
+        _ => event.localPosition,
+      };
+      _velocityTrackers[event.pointer]!.addPosition(event.timeStamp, position);
     }
     if (event is PointerMoveEvent && event.buttons != _initialButtons) {
       _giveUpPointer(event.pointer);
       return;
     }
-    if (event is PointerMoveEvent || event is PointerPanZoomUpdateEvent) {
-      final Offset delta = (event is PointerMoveEvent)
-          ? event.delta
-          : (event as PointerPanZoomUpdateEvent).panDelta;
-      final Offset localDelta = (event is PointerMoveEvent)
-          ? event.localDelta
-          : (event as PointerPanZoomUpdateEvent).localPanDelta;
-      final Offset position = (event is PointerMoveEvent)
-          ? event.position
-          : (event.position + (event as PointerPanZoomUpdateEvent).pan);
-      final Offset localPosition = (event is PointerMoveEvent)
-          ? event.localPosition
-          : (event.localPosition +
-              (event as PointerPanZoomUpdateEvent).localPan);
-      if (_state == _DragState.accepted) {
-        _checkUpdate(
-          sourceTimeStamp: event.timeStamp,
-          delta: _getDeltaForDetails(localDelta),
-          primaryDelta: _getPrimaryValueFromOffset(localDelta),
-          globalPosition: position,
-          localPosition: localPosition,
-        );
-      } else {
-        _pendingDragOffset += OffsetPair(local: localDelta, global: delta);
-        _lastPendingEventTimestamp = event.timeStamp;
-        _lastTransform = event.transform;
-        final Offset movedLocally = _getDeltaForDetails(localDelta);
-        final Matrix4? localToGlobalTransform = event.transform == null
-            ? null
-            : Matrix4.tryInvert(event.transform!);
-        _globalDistanceMoved += PointerEvent.transformDeltaViaPositions(
-                    transform: localToGlobalTransform,
-                    untransformedDelta: movedLocally,
-                    untransformedEndPosition: localPosition)
-                .distance *
-            (_getPrimaryValueFromOffset(movedLocally) ?? 1).sign;
-        if (_hasSufficientGlobalDistanceToAccept(
-            event.kind, gestureSettings?.touchSlop)) {
-          _hasDragThresholdBeenMet = true;
-          if (_acceptedActivePointers.contains(event.pointer)) {
-            _checkDrag(event.pointer);
-          } else {
-            resolve(GestureDisposition.accepted);
+    if ((event is PointerMoveEvent || event is PointerPanZoomUpdateEvent) &&
+        _shouldTrackMoveEvent(event.pointer)) {
+      final Offset delta =
+          (event is PointerMoveEvent)
+              ? event.delta
+              : (event as PointerPanZoomUpdateEvent).panDelta;
+      final Offset localDelta =
+          (event is PointerMoveEvent)
+              ? event.localDelta
+              : (event as PointerPanZoomUpdateEvent).localPanDelta;
+      final Offset position =
+          (event is PointerMoveEvent)
+              ? event.position
+              : (event.position + (event as PointerPanZoomUpdateEvent).pan);
+      final Offset localPosition =
+          (event is PointerMoveEvent)
+              ? event.localPosition
+              : (event.localPosition +
+                  (event as PointerPanZoomUpdateEvent).localPan);
+      _lastPosition = OffsetPair(local: localPosition, global: position);
+      final Offset resolvedDelta = _resolveLocalDeltaForMultitouch(
+        event.pointer,
+        localDelta,
+      );
+      switch (_state) {
+        case _DragState.ready || _DragState.possible:
+          _pendingDragOffset += OffsetPair(local: localDelta, global: delta);
+          _lastPendingEventTimestamp = event.timeStamp;
+          _lastTransform = event.transform;
+          final Offset movedLocally = _getDeltaForDetails(localDelta);
+          final Matrix4? localToGlobalTransform =
+              event.transform == null
+                  ? null
+                  : Matrix4.tryInvert(event.transform!);
+          _globalDistanceMoved +=
+              PointerEvent.transformDeltaViaPositions(
+                transform: localToGlobalTransform,
+                untransformedDelta: movedLocally,
+                untransformedEndPosition: localPosition,
+              ).distance *
+              (_getPrimaryValueFromOffset(movedLocally) ?? 1).sign;
+          if (hasSufficientGlobalDistanceToAccept(
+            event.kind,
+            gestureSettings?.touchSlop,
+          )) {
+            _hasDragThresholdBeenMet = true;
+            if (_acceptedActivePointers.contains(event.pointer)) {
+              _checkDrag(event.pointer);
+            } else {
+              resolve(GestureDisposition.accepted);
+            }
           }
-        }
+        case _DragState.accepted:
+          _checkUpdate(
+            sourceTimeStamp: event.timeStamp,
+            delta: _getDeltaForDetails(resolvedDelta),
+            primaryDelta: _getPrimaryValueFromOffset(resolvedDelta),
+            globalPosition: position,
+            localPosition: localPosition,
+          );
       }
+      _recordMoveDeltaForMultitouch(event.pointer, localDelta);
     }
-    if (event is PointerUpEvent ||
-        event is PointerCancelEvent ||
-        event is PointerPanZoomEndEvent) {
+    if (event
+        case PointerUpEvent() ||
+            PointerCancelEvent() ||
+            PointerPanZoomEndEvent()) {
       _giveUpPointer(event.pointer);
     }
   }
 
-  final Set<int> _acceptedActivePointers = <int>{};
+  final List<int> _acceptedActivePointers = <int>[];
+  // This value is used when the multitouch strategy is `latestPointer`,
+  // it keeps track of the last accepted pointer. If this active pointer
+  // leave up, it will be set to the first accepted pointer.
+  // Refer to the implementation of Android `RecyclerView`(line 3846):
+  // https://android.googlesource.com/platform/frameworks/support/+/refs/heads/androidx-main/recyclerview/recyclerview/src/main/java/androidx/recyclerview/widget/RecyclerView.java
+  int? _activePointer;
 
   @override
   void acceptGesture(int pointer) {
     assert(!_acceptedActivePointers.contains(pointer));
     _acceptedActivePointers.add(pointer);
+    _activePointer = pointer;
     if (!onlyAcceptDragOnThreshold || _hasDragThresholdBeenMet) {
       _checkDrag(pointer);
     }
@@ -434,10 +763,9 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
       case _DragState.possible:
         resolve(GestureDisposition.rejected);
         _checkCancel();
-        break;
+
       case _DragState.accepted:
         _checkEnd(pointer);
-        break;
     }
     _hasDragThresholdBeenMet = false;
     _velocityTrackers.clear();
@@ -451,6 +779,14 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
     // interested in winning the gesture arena for it.
     if (!_acceptedActivePointers.remove(pointer)) {
       resolvePointer(pointer, GestureDisposition.rejected);
+    }
+
+    _moveDeltaBeforeFrame.remove(pointer);
+    if (_activePointer == pointer) {
+      _activePointer =
+          _acceptedActivePointers.isNotEmpty
+              ? _acceptedActivePointers.first
+              : null;
     }
   }
 
@@ -477,10 +813,8 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
       case DragStartBehavior.start:
         _initialPosition = _initialPosition + delta;
         localUpdateDelta = Offset.zero;
-        break;
       case DragStartBehavior.down:
         localUpdateDelta = _getDeltaForDetails(delta.local);
-        break;
     }
     _pendingDragOffset = OffsetPair.zero;
     _lastPendingEventTimestamp = null;
@@ -496,8 +830,10 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
         untransformedDelta: localUpdateDelta,
         transform: localToGlobal,
       );
-      final OffsetPair updateDelta =
-          OffsetPair(local: localUpdateDelta, global: globalUpdateDelta);
+      final OffsetPair updateDelta = OffsetPair(
+        local: localUpdateDelta,
+        global: globalUpdateDelta,
+      );
       final OffsetPair correctedPosition =
           _initialPosition + updateDelta; // Only adds delta for down behaviour
       _checkUpdate(
@@ -558,15 +894,23 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
     if (estimate == null) {
       debugReport = () => 'Could not estimate velocity.';
     } else {
-      details = _considerFling(estimate, tracker.kind);
-      debugReport = (details != null)
-          ? () => '$estimate; fling at ${details!.velocity}.'
-          : () => '$estimate; judged to not be a fling.';
+      details = considerFling(estimate, tracker.kind);
+      debugReport =
+          (details != null)
+              ? () => '$estimate; fling at ${details!.velocity}.'
+              : () => '$estimate; judged to not be a fling.';
     }
-    details ??= DragEndDetails(primaryVelocity: 0.0);
+    details ??= DragEndDetails(
+      primaryVelocity: 0.0,
+      globalPosition: _lastPosition.global,
+      localPosition: _lastPosition.local,
+    );
 
-    invokeCallback<void>('onEnd', () => onEnd!(details!),
-        debugReport: debugReport);
+    invokeCallback<void>(
+      'onEnd',
+      () => onEnd!(details!),
+      debugReport: debugReport,
+    );
   }
 
   void _checkCancel() {
@@ -585,7 +929,8 @@ abstract class _DragGestureRecognizer extends OneSequenceGestureRecognizer {
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
     properties.add(
-        EnumProperty<DragStartBehavior>('start behavior', dragStartBehavior));
+      EnumProperty<DragStartBehavior>('start behavior', dragStartBehavior),
+    );
   }
 }
 
@@ -619,24 +964,33 @@ class _VerticalDragGestureRecognizer extends _DragGestureRecognizer {
   }
 
   @override
-  DragEndDetails? _considerFling(
-      VelocityEstimate estimate, PointerDeviceKind kind) {
+  DragEndDetails? considerFling(
+    VelocityEstimate estimate,
+    PointerDeviceKind kind,
+  ) {
     if (!isFlingGesture(estimate, kind)) {
       return null;
     }
     final double maxVelocity = maxFlingVelocity ?? kMaxFlingVelocity;
-    final double dy =
-        clampDouble(estimate.pixelsPerSecond.dy, -maxVelocity, maxVelocity);
+    final double dy = clampDouble(
+      estimate.pixelsPerSecond.dy,
+      -maxVelocity,
+      maxVelocity,
+    );
     return DragEndDetails(
       velocity: Velocity(pixelsPerSecond: Offset(0, dy)),
       primaryVelocity: dy,
+      globalPosition: lastPosition.global,
+      localPosition: lastPosition.local,
     );
   }
 
   @override
-  bool _hasSufficientGlobalDistanceToAccept(
-      PointerDeviceKind pointerDeviceKind, double? deviceTouchSlop) {
-    return _globalDistanceMoved.abs() >
+  bool hasSufficientGlobalDistanceToAccept(
+    PointerDeviceKind pointerDeviceKind,
+    double? deviceTouchSlop,
+  ) {
+    return globalDistanceMoved.abs() >
         computeHitSlop(pointerDeviceKind, gestureSettings);
   }
 
@@ -645,6 +999,9 @@ class _VerticalDragGestureRecognizer extends _DragGestureRecognizer {
 
   @override
   double _getPrimaryValueFromOffset(Offset value) => value.dy;
+
+  @override
+  _DragDirection? _getPrimaryDragAxis() => _DragDirection.vertical;
 
   @override
   String get debugDescription => 'vertical drag';
@@ -680,24 +1037,33 @@ class _HorizontalDragGestureRecognizer extends _DragGestureRecognizer {
   }
 
   @override
-  DragEndDetails? _considerFling(
-      VelocityEstimate estimate, PointerDeviceKind kind) {
+  DragEndDetails? considerFling(
+    VelocityEstimate estimate,
+    PointerDeviceKind kind,
+  ) {
     if (!isFlingGesture(estimate, kind)) {
       return null;
     }
     final double maxVelocity = maxFlingVelocity ?? kMaxFlingVelocity;
-    final double dx =
-        clampDouble(estimate.pixelsPerSecond.dx, -maxVelocity, maxVelocity);
+    final double dx = clampDouble(
+      estimate.pixelsPerSecond.dx,
+      -maxVelocity,
+      maxVelocity,
+    );
     return DragEndDetails(
       velocity: Velocity(pixelsPerSecond: Offset(dx, 0)),
       primaryVelocity: dx,
+      globalPosition: _lastPosition.global,
+      localPosition: _lastPosition.local,
     );
   }
 
   @override
-  bool _hasSufficientGlobalDistanceToAccept(
-      PointerDeviceKind pointerDeviceKind, double? deviceTouchSlop) {
-    return _globalDistanceMoved.abs() >
+  bool hasSufficientGlobalDistanceToAccept(
+    PointerDeviceKind pointerDeviceKind,
+    double? deviceTouchSlop,
+  ) {
+    return globalDistanceMoved.abs() >
         computeHitSlop(pointerDeviceKind, gestureSettings);
   }
 
@@ -708,168 +1074,8 @@ class _HorizontalDragGestureRecognizer extends _DragGestureRecognizer {
   double _getPrimaryValueFromOffset(Offset value) => value.dx;
 
   @override
+  _DragDirection? _getPrimaryDragAxis() => _DragDirection.horizontal;
+
+  @override
   String get debugDescription => 'horizontal drag';
-}
-
-class _PointAtTime {
-  const _PointAtTime(this.point, this.time);
-
-  final Duration time;
-  final Offset point;
-
-  @override
-  String toString() => '_PointAtTime($point at $time)';
-}
-
-// Computes a pointer's velocity based on data from [PointerMoveEvent]s.
-///
-/// The input data is provided by calling [addPosition]. Adding data is cheap.
-///
-/// To obtain a velocity, call [getVelocity] or [getVelocityEstimate]. This will
-/// compute the velocity based on the data added so far. Only call these when
-/// you need to use the velocity, as they are comparatively expensive.
-///
-/// The quality of the velocity estimation will be better if more data points
-/// have been received.
-class _VelocityTracker extends VelocityTracker {
-  /// Create a new velocity tracker for a pointer [kind].
-  _VelocityTracker.withKind(this.kind) : super.withKind(kind);
-
-  static const int _assumePointerMoveStoppedMilliseconds = 40;
-  static const int _historySize = 20;
-  static const int _horizonMilliseconds = 100;
-  static const int _minSampleSize = 3;
-
-  /// The kind of pointer this tracker is for.
-  @override
-  final PointerDeviceKind kind;
-
-  // Time difference since the last sample was added
-  final Stopwatch _sinceLastSample = Stopwatch();
-
-  // Circular buffer; current sample at _index.
-  final List<_PointAtTime?> _samples =
-      List<_PointAtTime?>.filled(_historySize, null);
-  int _index = 0;
-
-  /// Adds a position as the given time to the tracker.
-  @override
-  void addPosition(Duration time, Offset position) {
-    _sinceLastSample.start();
-    _sinceLastSample.reset();
-    _index += 1;
-    if (_index == _historySize) {
-      _index = 0;
-    }
-    _samples[_index] = _PointAtTime(position, time);
-  }
-
-  /// Returns an estimate of the velocity of the object being tracked by the
-  /// tracker given the current information available to the tracker.
-  ///
-  /// Information is added using [addPosition].
-  ///
-  /// Returns null if there is no data on which to base an estimate.
-  @override
-  VelocityEstimate? getVelocityEstimate() {
-    // no recent user movement?
-    if (_sinceLastSample.elapsedMilliseconds >
-        _VelocityTracker._assumePointerMoveStoppedMilliseconds) {
-      return const VelocityEstimate(
-        pixelsPerSecond: Offset.zero,
-        confidence: 1.0,
-        duration: Duration.zero,
-        offset: Offset.zero,
-      );
-    }
-
-    final List<double> x = <double>[];
-    final List<double> y = <double>[];
-    final List<double> w = <double>[];
-    final List<double> time = <double>[];
-    int sampleCount = 0;
-    int index = _index;
-
-    final _PointAtTime? newestSample = _samples[index];
-    if (newestSample == null) {
-      return null;
-    }
-
-    _PointAtTime previousSample = newestSample;
-    _PointAtTime oldestSample = newestSample;
-
-    // Starting with the most recent PointAtTime sample, iterate backwards while
-    // the samples represent continuous motion.
-    do {
-      final _PointAtTime? sample = _samples[index];
-      if (sample == null) {
-        break;
-      }
-
-      final double age =
-          (newestSample.time - sample.time).inMicroseconds.toDouble() / 1000;
-      final double delta =
-          (sample.time - previousSample.time).inMicroseconds.abs().toDouble() /
-              1000;
-      previousSample = sample;
-      if (age > _horizonMilliseconds ||
-          delta > _VelocityTracker._assumePointerMoveStoppedMilliseconds) {
-        break;
-      }
-
-      oldestSample = sample;
-      final Offset position = sample.point;
-      x.add(position.dx);
-      y.add(position.dy);
-      w.add(1.0);
-      time.add(-age);
-      index = (index == 0 ? _historySize : index) - 1;
-
-      sampleCount += 1;
-    } while (sampleCount < _historySize);
-
-    if (sampleCount >= _minSampleSize) {
-      final LeastSquaresSolver xSolver = LeastSquaresSolver(time, x, w);
-      final PolynomialFit? xFit = xSolver.solve(2);
-      if (xFit != null) {
-        final LeastSquaresSolver ySolver = LeastSquaresSolver(time, y, w);
-        final PolynomialFit? yFit = ySolver.solve(2);
-        if (yFit != null) {
-          return VelocityEstimate(
-            // convert from pixels/ms to pixels/s
-            pixelsPerSecond: Offset(
-                xFit.coefficients[1] * 1000, yFit.coefficients[1] * 1000),
-            confidence: xFit.confidence * yFit.confidence,
-            duration: newestSample.time - oldestSample.time,
-            offset: newestSample.point - oldestSample.point,
-          );
-        }
-      }
-    }
-
-    // We're unable to make a velocity estimate but we did have at least one
-    // valid pointer position.
-    return VelocityEstimate(
-      pixelsPerSecond: Offset.zero,
-      confidence: 1.0,
-      duration: newestSample.time - oldestSample.time,
-      offset: newestSample.point - oldestSample.point,
-    );
-  }
-
-  /// Computes the velocity of the pointer at the time of the last
-  /// provided data point.
-  ///
-  /// This can be expensive. Only call this when you need the velocity.
-  ///
-  /// Returns [Velocity.zero] if there is no data from which to compute an
-  /// estimate or if the estimated velocity is zero.
-  @override
-  Velocity getVelocity() {
-    final VelocityEstimate? estimate = getVelocityEstimate();
-    if (estimate == null || estimate.pixelsPerSecond == Offset.zero) {
-      return Velocity.zero;
-    }
-    return Velocity(pixelsPerSecond: estimate.pixelsPerSecond);
-  }
 }
