@@ -92,30 +92,31 @@ import 'package:flutter/material.dart';
 ///  * [DecorationImage], which holds a configuration for calling this function.
 ///  * [BoxDecoration], which uses this function to paint a [DecorationImage].
 
-void paintExtendedImage(
-    {required Canvas canvas,
-    required Rect rect,
-    required ui.Image image,
-    String? debugImageLabel,
-    double scale = 1.0,
-    double opacity = 1.0,
-    ColorFilter? colorFilter,
-    BoxFit? fit,
-    Alignment alignment = Alignment.center,
-    Rect? centerSlice,
-    ImageRepeat repeat = ImageRepeat.noRepeat,
-    bool flipHorizontally = false,
-    bool invertColors = false,
-    FilterQuality filterQuality = FilterQuality.low,
-    Rect? customSourceRect,
-    //you can paint anything if you want before paint image.
-    BeforePaintImage? beforePaintImage,
-    //you can paint anything if you want after paint image.
-    AfterPaintImage? afterPaintImage,
-    GestureDetails? gestureDetails,
-    EditActionDetails? editActionDetails,
-    bool isAntiAlias = false,
-    EdgeInsets layoutInsets = EdgeInsets.zero}) {
+void paintExtendedImage({
+  required Canvas canvas,
+  required Rect rect,
+  required ui.Image image,
+  String? debugImageLabel,
+  double scale = 1.0,
+  double opacity = 1.0,
+  ColorFilter? colorFilter,
+  BoxFit? fit,
+  Alignment alignment = Alignment.center,
+  Rect? centerSlice,
+  ImageRepeat repeat = ImageRepeat.noRepeat,
+  bool flipHorizontally = false,
+  bool invertColors = false,
+  FilterQuality filterQuality = FilterQuality.low,
+  Rect? customSourceRect,
+  //you can paint anything if you want before paint image.
+  BeforePaintImage? beforePaintImage,
+  //you can paint anything if you want after paint image.
+  AfterPaintImage? afterPaintImage,
+  GestureDetails? gestureDetails,
+  EditActionDetails? editActionDetails,
+  bool isAntiAlias = false,
+  EdgeInsets layoutInsets = EdgeInsets.zero,
+}) {
   assert(
     image.debugGetOpenHandleStackTraces()?.isNotEmpty ?? true,
     'Cannot paint an image that is disposed.\n'
@@ -144,15 +145,20 @@ void paintExtendedImage(
 
   Offset? sliceBorder;
   if (centerSlice != null) {
-    sliceBorder = Offset(centerSlice.left + inputSize.width - centerSlice.right,
-        centerSlice.top + inputSize.height - centerSlice.bottom);
+    sliceBorder = Offset(
+      centerSlice.left + inputSize.width - centerSlice.right,
+      centerSlice.top + inputSize.height - centerSlice.bottom,
+    );
     outputSize = outputSize - sliceBorder as Size;
     inputSize = inputSize - sliceBorder as Size;
   }
   fit ??= centerSlice == null ? BoxFit.scaleDown : BoxFit.fill;
   assert(centerSlice == null || (fit != BoxFit.none && fit != BoxFit.cover));
-  final FittedSizes fittedSizes =
-      applyBoxFit(fit, inputSize / scale, outputSize);
+  final FittedSizes fittedSizes = applyBoxFit(
+    fit,
+    inputSize / scale,
+    outputSize,
+  );
   final Size sourceSize = fittedSizes.source * scale;
   Size destinationSize = fittedSizes.destination;
   if (centerSlice != null) {
@@ -160,8 +166,10 @@ void paintExtendedImage(
     destinationSize += sliceBorder;
     // We don't have the ability to draw a subset of the image at the same time
     // as we apply a nine-patch stretch.
-    assert(sourceSize == inputSize,
-        'centerSlice was used with a BoxFit that does not guarantee that the image is fully visible.');
+    assert(
+      sourceSize == inputSize,
+      'centerSlice was used with a BoxFit that does not guarantee that the image is fully visible.',
+    );
   }
   if (repeat != ImageRepeat.noRepeat && destinationSize == outputSize) {
     // There's no need to repeat the image because we're exactly filling the
@@ -179,7 +187,8 @@ void paintExtendedImage(
       (outputSize.width - destinationSize.width) / 2.0;
   final double halfHeightDelta =
       (outputSize.height - destinationSize.height) / 2.0;
-  final double dx = halfWidthDelta +
+  final double dx =
+      halfWidthDelta +
       (flipHorizontally ? -alignment.x : alignment.x) * halfWidthDelta;
   final double dy = halfHeightDelta + alignment.y * halfHeightDelta;
   final Offset destinationPosition = topLeft.translate(dx, dy);
@@ -188,8 +197,10 @@ void paintExtendedImage(
   bool needClip = false;
 
   if (gestureDetails != null) {
-    destinationRect =
-        gestureDetails.calculateFinalDestinationRect(rect, destinationRect);
+    destinationRect = gestureDetails.calculateFinalDestinationRect(
+      rect,
+      destinationRect,
+    );
 
     // outside and need clip
     needClip = !rect.containsRect(destinationRect);
@@ -208,13 +219,14 @@ void paintExtendedImage(
   if (editActionDetails != null) {
     if (editActionDetails.cropRectPadding != null) {
       destinationRect = getDestinationRect(
-          inputSize: inputSize,
-          rect: editActionDetails.cropRectPadding!.deflateRect(rect),
-          fit: fit,
-          flipHorizontally: false,
-          scale: scale,
-          centerSlice: centerSlice,
-          alignment: alignment);
+        inputSize: inputSize,
+        rect: editActionDetails.cropRectPadding!.deflateRect(rect),
+        fit: fit,
+        flipHorizontally: false,
+        scale: scale,
+        centerSlice: centerSlice,
+        alignment: alignment,
+      );
     }
 
     editActionDetails.initRect(rect, destinationRect);
@@ -261,25 +273,41 @@ void paintExtendedImage(
   }
 
   if (centerSlice == null) {
-    final Rect sourceRect = customSourceRect ??
+    final Rect sourceRect =
+        customSourceRect ??
         alignment.inscribe(sourceSize, Offset.zero & inputSize);
     if (repeat == ImageRepeat.noRepeat) {
       canvas.drawImageRect(image, sourceRect, destinationRect, paint);
     } else {
-      for (final Rect tileRect
-          in _generateImageTileRects(rect, destinationRect, repeat))
+      for (final Rect tileRect in _generateImageTileRects(
+        rect,
+        destinationRect,
+        repeat,
+      )) {
         canvas.drawImageRect(image, sourceRect, tileRect, paint);
+      }
     }
   } else {
     canvas.scale(1 / scale);
     if (repeat == ImageRepeat.noRepeat) {
-      canvas.drawImageNine(image, _scaleRect(centerSlice, scale),
-          _scaleRect(destinationRect, scale), paint);
+      canvas.drawImageNine(
+        image,
+        _scaleRect(centerSlice, scale),
+        _scaleRect(destinationRect, scale),
+        paint,
+      );
     } else {
-      for (final Rect tileRect
-          in _generateImageTileRects(rect, destinationRect, repeat))
-        canvas.drawImageNine(image, _scaleRect(centerSlice, scale),
-            _scaleRect(tileRect, scale), paint);
+      for (final Rect tileRect in _generateImageTileRects(
+        rect,
+        destinationRect,
+        repeat,
+      ))
+        canvas.drawImageNine(
+          image,
+          _scaleRect(centerSlice, scale),
+          _scaleRect(tileRect, scale),
+          paint,
+        );
     }
   }
 
@@ -306,7 +334,10 @@ void paintExtendedImage(
 }
 
 List<Rect> _generateImageTileRects(
-    Rect outputRect, Rect fundamentalRect, ImageRepeat repeat) {
+  Rect outputRect,
+  Rect fundamentalRect,
+  ImageRepeat repeat,
+) {
   int startX = 0;
   int startY = 0;
   int stopX = 0;
@@ -331,5 +362,9 @@ List<Rect> _generateImageTileRects(
   ];
 }
 
-Rect _scaleRect(Rect rect, double scale) => Rect.fromLTRB(rect.left * scale,
-    rect.top * scale, rect.right * scale, rect.bottom * scale);
+Rect _scaleRect(Rect rect, double scale) => Rect.fromLTRB(
+  rect.left * scale,
+  rect.top * scale,
+  rect.right * scale,
+  rect.bottom * scale,
+);
