@@ -17,12 +17,12 @@ typedef LoadStateChanged = Widget? Function(ExtendedImageState state);
 
 /// [rect] is render size
 /// if return true, it will not paint original image,
-typedef BeforePaintImage = bool Function(
-    Canvas canvas, Rect rect, ui.Image image, Paint paint);
+typedef BeforePaintImage =
+    bool Function(Canvas canvas, Rect rect, ui.Image image, Paint paint);
 
 /// Call after paint image
-typedef AfterPaintImage = void Function(
-    Canvas canvas, Rect rect, ui.Image image, Paint paint);
+typedef AfterPaintImage =
+    void Function(Canvas canvas, Rect rect, ui.Image image, Paint paint);
 
 /// Animation call back for inertia drag
 typedef GestureOffsetAnimationCallBack = void Function(Offset offset);
@@ -34,32 +34,29 @@ typedef GestureScaleAnimationCallBack = void Function(double scale);
 typedef DoubleTap = void Function(ExtendedImageGestureState state);
 
 /// Build page background when slide page
-typedef SlidePageBackgroundHandler = Color Function(
-    Offset offset, Size pageSize);
+typedef SlidePageBackgroundHandler =
+    Color Function(Offset offset, Size pageSize);
 
 /// customize offset of page when slide page
-typedef SlideOffsetHandler = Offset? Function(
-  Offset offset, {
-  ExtendedImageSlidePageState state,
-});
+typedef SlideOffsetHandler =
+    Offset? Function(Offset offset, {ExtendedImageSlidePageState state});
 
 /// if return true ,pop page
 /// else reset page state
-typedef SlideEndHandler = bool? Function(
-  Offset offset, {
-  ExtendedImageSlidePageState state,
-  ScaleEndDetails details,
-});
+typedef SlideEndHandler =
+    bool? Function(
+      Offset offset, {
+      ExtendedImageSlidePageState state,
+      ScaleEndDetails details,
+    });
 
 /// Customize scale of page when slide page
-typedef SlideScaleHandler = double? Function(
-  Offset offset, {
-  ExtendedImageSlidePageState state,
-});
+typedef SlideScaleHandler =
+    double? Function(Offset offset, {ExtendedImageSlidePageState state});
 
 /// Init GestureConfig when image is ready.
-typedef InitGestureConfigHandler = GestureConfig Function(
-    ExtendedImageState state);
+typedef InitGestureConfigHandler =
+    GestureConfig Function(ExtendedImageState state);
 
 /// Call on sliding page
 typedef OnSlidingPage = void Function(ExtendedImageSlidePageState state);
@@ -77,12 +74,12 @@ typedef MergeEditRect = Rect Function(Rect editRect);
 typedef BuildGestureImage = Widget Function(GestureDetails gestureDetails);
 
 /// Init GestureConfig when image is ready.
-typedef InitEditorConfigHandler = EditorConfig? Function(
-    ExtendedImageState? state);
+typedef InitEditorConfigHandler =
+    EditorConfig? Function(ExtendedImageState? state);
 
 /// Get editor mask color base on pointerDown
-typedef EditorMaskColorHandler = Color Function(
-    BuildContext context, bool pointerDown);
+typedef EditorMaskColorHandler =
+    Color Function(BuildContext context, bool pointerDown);
 
 /// Build Hero only for sliding page
 /// the transform of sliding page must be working on Hero
@@ -90,10 +87,11 @@ typedef EditorMaskColorHandler = Color Function(
 typedef HeroBuilderForSlidingPage = Widget Function(Widget widget);
 
 /// Build image for gesture, we can handle custom Widget about gesture
-typedef ImageBuilderForGesture = Widget Function(
-  Widget image, {
-  ExtendedImageGestureState? imageGestureState,
-});
+typedef ImageBuilderForGesture =
+    Widget Function(
+      Widget image, {
+      ExtendedImageGestureState? imageGestureState,
+    });
 
 /// Whether should scale image
 typedef CanScaleImage = bool Function(GestureDetails? details);
